fix: check server responds to HTTP requests before allowing user messages
