+++ browser-tests/tests/traces.test.ts
@@ -1,19 +1,10 @@
-import { expect, test } from '../fixtures';
+import { expect, sendMessage, test } from '../fixtures';
 
 test.describe('Traces', () => {
   test('traces panel shows trace after invocation', async ({ page }) => {
     await page.goto('/');
 
-    const chatInput = page.getByTestId('chat-input');
-    await expect(chatInput).toBeEnabled({ timeout: 60_000 });
-
-    await chatInput.fill('Say hello');
-    await page.getByRole('button', { name: 'Send message' }).click();
-
-    const messageList = page.getByTestId('message-list');
-    const assistantMessage = messageList.getByTestId('chat-message-1');
-    await expect(assistantMessage).toBeVisible({ timeout: 60_000 });
-    await expect(assistantMessage).not.toContainText('ECONNREFUSED');
+    await sendMessage(page, 'Say hello');
 
     await page.getByRole('tab', { name: 'Traces' }).click();
